Comments for the new classes

+++ arena2d-sim/scripts/evaluation.py
@@ -37,6 +37,8 @@ time_out_counter = np.count_nonzero(ending == 'time')
 goal_counter = np.count_nonzero(ending == 'goal')
 num_episodes = human_counter + wall_counter + time_out_counter + goal_counter
 
+#check if there were humans in level
+human_exist = human_robot_distance.size != 0
 
 #read safty distance and time-out time from settings.st
 with open("settings.st", "r") as f:
@@ -56,21 +58,8 @@ for line in settings:
 		end = line.find("#")
 		time_step = float(line[start+1:end])
 
-	if 'goal_size = ' in line:
-		start = line.find("=")
-		end = line.find("#")
-		goal_size = float(line[start+1:end])
-	if base:
-		start = line.find("=")
-		base_size = float(line[start+1:])
-		base = False
-	if 'base_size' in line:
-		base = True
-
 		
 max_step_time_out = max_time/time_step
-#correct direkt goal distance
-goal_distance = goal_distance - (goal_size/2 + base_size)
 
 #get time_to_goal and traveled distanse per episode
 time_to_goal = []
@@ -81,9 +70,15 @@ for current_episode in range(num_episodes):
 	idx_start = np.argwhere(episode == current_episode) if current_episode != 0 else 0
 	idx_end = np.argwhere(episode == current_episode+1)
 	if ending[idx_end] == 'goal':
+		#correct robot position index mistake if episode_end != time_out
+		if ending[idx_start] != 'time':
+			idx_start = idx_start + 2 #if last episode was time then correction is NOT nessesary 
+		idx_end = idx_end + 2 #must always be corrected if ending[idx_end] != time
+			
 		position_for_episode = robot_position[int(idx_start):int(idx_end)]
 		position_for_episode = position_for_episode[~np.isnan(position_for_episode).any(axis=1)]
-		time_to_goal.append(position_for_episode.shape[0])
+
+		time_to_goal.append(position_for_episode.shape[0] - 1)
 		dist = 0
 		for row_idx in range(position_for_episode.shape[0]-1):
 			v1 = position_for_episode[row_idx]
@@ -92,52 +87,32 @@ for current_episode in range(num_episodes):
 		traveled_distance.append(dist)
 		#first episode goal info missing
 		if current_episode != 0:
-			frac_direct_traveled_dist.append(dist/goal_distance[current_episode-1]) 
-			if dist/goal_distance[current_episode-1]<1:
-				temp = temp+1
-print(temp)
+			frac_direct_traveled_dist.append(dist/goal_distance[current_episode-1])
+
+
 
 #remove all nan 
 robot_position = robot_position[~np.isnan(robot_position).any(axis=1)]
 robot_direction = robot_direction[~np.isnan(robot_direction).any(axis=1)]
 time_to_goal = np.array(time_to_goal)
-#traveled_distance = np.array(traveled_distance)
-
-#print hit numbers
-#print('for evaluation the robot did 1000 episodes')
-#print('Robot hit ' + str(human_counter) + ' times a human')
-#print('Robot hit ' + str(wall_counter) + ' times a wall')
-#print('Robot reaches ' + str(time_out_counter) + ' times NOT the goal')
-
-text_file = open(path + "evaluation_stats.txt", "w")
-text_file.write('For the evaluation the robot did ' + str(num_episodes) + ' episodes \n')
-text_file.write('The robot reached ' + str(goal_counter) + ' the goal \n')
-text_file.write('The robot hit ' + str(human_counter) + ' times a human \n')
-text_file.write('The robot hit ' + str(wall_counter) + ' times a wall \n')
-text_file.write('The robot didn\'t reach the goal in time for ' + str(time_out_counter) + ' times \n')
-text_file.write('---------------------------------------------------------------- \n')
-text_file.write('Success rate: ' + str(round(goal_counter/num_episodes*100,1)) + '%\n')
-text_file.write('Human hit rate: ' + str(round(human_counter/num_episodes*100,1)) + '%\n')
-text_file.write('Wall hit rate: ' + str(round(wall_counter/num_episodes*100,1)) + '%\n')
-text_file.write('Timeout rate: ' + str(round(time_out_counter/num_episodes*100,1)) + '%\n')
-text_file.close()
 
 
 
 #hist plot of distances robot-human
-mu = np.mean(human_robot_distance)
-std=np.std(human_robot_distance)
+if human_exist:
+	mu = np.mean(human_robot_distance)
+	std=np.std(human_robot_distance)
 
-plt.figure(1)
-plt.hist(human_robot_distance.flatten(),bins=50,align = 'right',edgecolor='black', linewidth=0.8)
-plt.axvline(safety_distance_human, color='r', linestyle='dashed', linewidth=1.5)
-plt.annotate('safety distance = %.2f'%(safety_distance_human), xy=(0, 0.85), xycoords='axes fraction',color = 'red')
-plt.annotate('$\mu = %.2f$'%(mu) + ' \n$\sigma^2 = %.2f$'%(std), xy=(0.85, 0.85), xycoords='axes fraction',bbox=dict(boxstyle="round", fc="w"))
-plt.title('Hist of human distances')
-plt.xlabel('distance')
-plt.ylabel('counts')
-plt.savefig(path+ 'human_distance_hist')
-plt.clf()
+	plt.figure(1)
+	plt.hist(human_robot_distance.flatten(),bins=50,align = 'right',edgecolor='black', linewidth=0.8)
+	plt.axvline(safety_distance_human, color='r', linestyle='dashed', linewidth=1.5)
+	plt.annotate('safety distance = %.2f'%(safety_distance_human), xy=(0, 0.85), xycoords='axes fraction',color = 'red')
+	plt.annotate('$\mu = %.2f$'%(mu) + ' \n$\sigma^2 = %.2f$'%(std), xy=(0.85, 0.85), xycoords='axes fraction',bbox=dict(boxstyle="round", fc="w"))
+	plt.title('Hist of human distances')
+	plt.xlabel('distance')
+	plt.ylabel('counts')
+	plt.savefig(path+ 'human_distance_hist')
+	plt.clf()
 
 #box plot auf time to reach goal 
 plt.figure(2)
@@ -164,5 +139,55 @@ plt.boxplot(frac_direct_traveled_dist)
 plt.savefig(path + 'goal_distance_box')
 plt.clf()
 
-print('evaluation done')
+#calculate and plot ending counters per bins and in total and write text file
+bin_size = 100
+ending = ending[~pd.isnull(data['Ending'])] #remove all nan
+ending = ending[ending != 'reset'] # remove all reset
+ending = np.reshape(ending,(-1,bin_size))# reshape in bins of 100
 
+wall_counter_bin = []
+human_counter_bin = []
+time_out_counter_bin = []
+goal_counter_bin = []
+
+for row in ending:
+	wall_counter_bin.append(np.count_nonzero(row == 'wall')/bin_size*100)
+	human_counter_bin.append(np.count_nonzero(row == 'human')/bin_size*100)
+	time_out_counter_bin.append(np.count_nonzero(row == 'time')/bin_size*100)
+	goal_counter_bin.append(np.count_nonzero(row == 'goal')/bin_size*100)
+
+#write text file
+text_file = open(path + "evaluation_stats.txt", "w")
+text_file.write('For the evaluation the robot did ' + str(num_episodes) + ' episodes \n')
+text_file.write('The robot reached ' + str(goal_counter) + ' the goal \n')
+if human_exist:
+	text_file.write('The robot hit ' + str(human_counter) + ' times a human \n')
+else:
+	text_file.write('no human in level \n')
+text_file.write('The robot hit ' + str(wall_counter) + ' times a wall \n')
+text_file.write('The robot didn\'t reach the goal in time for ' + str(time_out_counter) + ' times \n')
+text_file.write('---------------------------------------------------------------- \n')
+text_file.write('Success rate: ' + str(round(goal_counter/num_episodes*100,1)) + '%')
+text_file.write(', Variance: ' + str(round(np.std(goal_counter_bin),1)) + '%\n')
+text_file.write('Human hit rate: ' + str(round(human_counter/num_episodes*100,1)) + '%')
+text_file.write(', Variance: ' + str(round(np.std(human_counter_bin),1)) + '%\n')
+text_file.write('Wall hit rate: ' + str(round(wall_counter/num_episodes*100,1)) + '%')
+text_file.write(', Variance: ' + str(round(np.std(wall_counter_bin),1)) + '%\n')
+text_file.write('Timeout rate: ' + str(round(time_out_counter/num_episodes*100,1)) + '%')
+text_file.write(', Variance: ' + str(round(np.std(time_out_counter_bin),1)) + '%\n')
+text_file.close()
+#plot
+plt.figure(4)
+plt.title('episode endings over bins with size ' + str(bin_size))
+plt.ylabel('number of counted endings')
+plt.xlabel('bin')
+plt.plot(wall_counter_bin,label='wall')
+plt.plot(human_counter_bin,label='human')
+plt.plot(time_out_counter_bin,label='time out')
+plt.plot(goal_counter_bin,label='goal')
+plt.legend()
+plt.savefig(path + 'endings')
+plt.clf()
+
+
+print('evaluation done')
